requirements.txt
PyQt5
pywin32
qtutils
requests
bs4
qtpy
qrcode
pillow
nuitka